web/requirements.txt
Flask
flask_restful
pymongo
bcrypt
numpy
tensorflow
requests
python-dotenv
pydns